requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# FastAPI Web Interface
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# RAG & Vector Store
beautifulsoup4
lxml>=4.9.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0
